+++ tests/data/test_import_consortium.py
@@ -11,7 +11,8 @@ import pathlib
 
 from click.testing import CliRunner
 
-from src.data.import_consortium import main as import_consortium
+from src.data.import_consortium import (main as
+    import_consortium, check_chromosomes)
 from src.features.smarterdb import VariantSheep
 
 from ..common import MongoMockMixin, VariantsMixin
@@ -65,3 +66,8 @@ class ImportConsortiumTest(VariantsMixin, MongoMockMixin, unittest.TestCase):
         self.assertEqual(location.chrom, "15")
         self.assertEqual(location.position, 5870057)
         self.assertEqual(location.illumina_top, "T/C")
+
+    def test_check_chromosomes(self):
+        self.assertEqual(check_chromosomes("26", "Sheep"), "26")
+        self.assertEqual(check_chromosomes("27", "Sheep"), "X")
+        self.assertRaises(NotImplementedError, check_chromosomes, 1, "Goat")